+++ platforms/bk7231t/bk7231t_os/application.mk
@@ -192,7 +192,7 @@ SRC_C += ./beken378/driver/common/dd.c
 SRC_C += ./beken378/driver/common/drv_model.c
 SRC_C += ./beken378/driver/dma/dma.c
 SRC_C += ./beken378/driver/driver.c
-SRC_CPP += ./beken378/driver/entry/arch_main.cpp
+SRC_C += ./beken378/driver/entry/arch_main.c
 SRC_C += ./beken378/driver/fft/fft.c
 SRC_C += ./beken378/driver/flash/flash.c
 SRC_C += ./beken378/driver/general_dma/general_dma.c
